FlaskRecap/requirements.txt
aniso8601==6.0.0
bcrypt==3.2.0
cffi==1.14.6
Click==7.0
ecdsa==0.17.0
Flask==1.0.3
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.2
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.18.0
python-jose==3.3.0
pytz==2019.1
rsa==4.7.2
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==0.15.5
